# openapi.yaml
openapi: 3.0.3
info:
  title: Purrfect API
  version: 1.0.0
  description: >
    Secure cat marketplace backend with escrow milestone settlement,
    veterinary inspection window, and dispute moderation.
servers:
  - url: https://api.purrfect.kz
    description: Production
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /orders:
    post:
      tags: [orders]
      summary: Create order and fund escrow (100%)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order funded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [orders]
      summary: List my orders (cursor pagination)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [CREATED, FUNDED_100, HANDOVER_CONFIRMED, INSPECTION_ACTIVE, DISPUTED, COMPLETED, CANCELLED]
          example: INSPECTION_ACTIVE
      responses:
        '200':
          description: Orders page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrders'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /orders/{orderId}:
    get:
      tags: [orders]
      summary: Get order detail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /admin/users/{userId}/role:
    patch:
      tags: [admin]
      summary: Update user role as admin
      description: Assigns a new RBAC role to a platform user. Requires ADMIN role.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: User role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/users/{userId}/status:
    patch:
      tags: [admin]
      summary: Update user status as admin
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: User status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/audit-logs:
    get:
      tags: [admin]
      summary: List audit logs as admin
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: action
          schema:
            type: string
      responses:
        '200':
          description: Audit logs page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogs'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /users/{userId}:
    get:
      tags: [users]
      summary: Public profile of any user (limited fields)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Public profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  displayName: { type: string }
                  role: { type: string, enum: [BUYER, SELLER, MODERATOR, ADMIN] }
                  trustScore: { type: integer }
                  createdAt: { type: string, format: date-time }
                  activeListings: { type: integer, example: 4 }
                  completedSales: { type: integer, example: 11 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /admin/dashboard-kpis:
    get:
      tags: [admin]
      summary: Get admin dashboard KPI summary
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: KPI summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  usersTotal: { type: integer, example: 2400 }
                  listingsPublished: { type: integer, example: 312 }
                  disputesOpen: { type: integer, example: 8 }
                  ordersCompleted: { type: integer, example: 128 }
                  activeRiskFlags: { type: integer, example: 11 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /admin/users/{userId}/ban:
    patch:
      tags: [admin]
      summary: Ban a user (sets status to SUSPENDED, revokes sessions). Reason is required.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 5
                  maxLength: 500
                  example: 'Multiple verified fraud reports'
      responses:
        '200':
          description: User banned (status=SUSPENDED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/users/{userId}:
    get:
      tags: [admin]
      summary: Detailed view of a single user (admin) with usage stats
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: User detail with stats
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      stats:
                        type: object
                        properties:
                          listingsCount: { type: integer }
                          activeListingsCount: { type: integer }
                          ordersAsBuyer: { type: integer }
                          ordersAsSeller: { type: integer }
                          openDisputes: { type: integer }
                          activeSessions: { type: integer }
                          auditLogsCount: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [admin]
      summary: Soft-delete a user (anonymize, revoke sessions). Refuses while orders are in flight.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: reason
          schema: { type: string, maxLength: 500 }
          example: 'GDPR erasure request'
      responses:
        '200':
          description: User soft-deleted
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/orders/{orderId}:
    get:
      tags: [admin]
      summary: Detailed view of a single order (admin) including escrow, payouts, inspection, dispute
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Order with all related entities
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: BAD_REQUEST
            message: Invalid query parameter value
    Unauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Access token is invalid or expired
...
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [listingId]
      properties:
        listingId:
          type: string
          format: uuid
          example: b8f198e7-ea6b-4d4a-a5bc-faf9d32d6ca7
        idempotencyKey:
          type: string
          example: ord_01J9J4XX45QCNH7Y
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        listingId:
          type: string
          format: uuid
        buyerId:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [CREATED, FUNDED_100, HANDOVER_CONFIRMED, INSPECTION_ACTIVE, DISPUTED, COMPLETED, CANCELLED]
        totalAmountKzt:
          type: number
          format: double
          example: 250000
        payout1Kzt:
          type: number
          format: double
          example: 118750
        payout2Kzt:
          type: number
          format: double
          example: 118750
        inspectionDeadline:
          type: string
          format: date-time
...